Renamed PWManifest to manifest
