+++ src/manifest.hpp
