+++ src/parser/parser_v1.cpp
@@ -4,7 +4,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2021 Datadog, Inc.
 
-#include <PWManifest.h>
+#include <manifest.hpp>
 #include <exception.hpp>
 #include <log.hpp>
 #include <parameter.hpp>
